requirements.txt
beautifulsoup4
edge-tts
face_recognition
flask
homeassistant_api
lxml
mediapipe
openai
opencv-python
ping3
psutil
pyaudio
pygame
pyserial
pywifi
rapidocr_openvino
sherpa_onnx
soundfile
ultralytics
zai-sdk

#sudo apt install cmake libasound-dev portaudio19-dev libportaudio2 libportaudiocpp0